fix: move common to ui/common (#324)
